bugz killa

+++ admin/network.h
@@ -28,11 +28,11 @@ public:
     Network(QObject *parent);
     void connectToServer();
     const Clients& getClients() const;
-    QString getClientName(qint32 id) const;
+    QString getClientName(qint16 id) const;
     void sendLevelOne(qint16 dest, const QByteArray& data);
-    void sendLevelTwo(qint32 dest, ProtocolMode mode, const QByteArray& data);
-    void activateMode(qint32 client, ProtocolMode mode);
-    void deactivateMode(qint32 client, ProtocolMode mode);
+    void sendLevelTwo(qint16 dest, ProtocolMode mode, const QByteArray& data);
+    void activateMode(qint16 client, ProtocolMode mode);
+    void deactivateMode(qint16 client, ProtocolMode mode);
     const Income& receivedData() const;
 signals:
     void listUpdated();
